refactor: simplify folder and code structure
