+++ docs/tryorama.conductorcell.cellid.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [ConductorCell](./tryorama.conductorcell.md) &gt; [cellId](./tryorama.conductorcell.cellid.md)
+
+## ConductorCell.cellId property
+
+**Signature:**
+
+```typescript
+cellId: CellId;
+```
