+++ docs/tryorama.conductorcell.conductor.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [ConductorCell](./tryorama.conductorcell.md) &gt; [conductor](./tryorama.conductorcell.conductor.md)
+
+## ConductorCell.conductor property
+
+**Signature:**
+
+```typescript
+conductor: Conductor;
+```
