+++ docs/tryorama.conductorcell.md
@@ -1,15 +1,15 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IConductorCell](./tryorama.iconductorcell.md)
+[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [ConductorCell](./tryorama.conductorcell.md)
 
-## IConductorCell interface
+## ConductorCell interface
 
 A Conductor and a CellId
 
 **Signature:**
 
 ```typescript
-export interface IConductorCell 
+export interface ConductorCell 
 ```
 
 ## Properties
@@ -37,7 +37,7 @@ Description
 </th></tr></thead>
 <tbody><tr><td>
 
-[cellId](./tryorama.iconductorcell.cellid.md)
+[cellId](./tryorama.conductorcell.cellid.md)
 
 
 </td><td>
@@ -54,7 +54,7 @@ CellId
 </td></tr>
 <tr><td>
 
-[conductor](./tryorama.iconductorcell.conductor.md)
+[conductor](./tryorama.conductorcell.conductor.md)
 
 
 </td><td>
@@ -62,7 +62,7 @@ CellId
 
 </td><td>
 
-[IConductor](./tryorama.iconductor.md)
+[Conductor](./tryorama.conductor.md)
 
 
 </td><td>
