--- docs/tryorama.iadminwebsocket.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IAdminWebsocket](./tryorama.iadminwebsocket.md)
-
-## IAdminWebsocket type
-
-AdminWebsocket interface for local and TryCP conductors.
-
-**Signature:**
-
-```typescript
-export type IAdminWebsocket = AdminWebsocket;
-```
