--- docs/tryorama.iappwebsocket.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IAppWebsocket](./tryorama.iappwebsocket.md)
-
-## IAppWebsocket type
-
-AppWebsocket interface for local and TryCP conductors.
-
-**Signature:**
-
-```typescript
-export type IAppWebsocket = AppWebsocket;
-```
