--- docs/tryorama.iconductor.adminws.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IConductor](./tryorama.iconductor.md) &gt; [adminWs](./tryorama.iconductor.adminws.md)
-
-## IConductor.adminWs property
-
-**Signature:**
-
-```typescript
-adminWs: () => IAdminWebsocket;
-```
