--- docs/tryorama.iconductor.connectappws.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IConductor](./tryorama.iconductor.md) &gt; [connectAppWs](./tryorama.iconductor.connectappws.md)
-
-## IConductor.connectAppWs property
-
-**Signature:**
-
-```typescript
-connectAppWs: (token: AppAuthenticationToken, port: number) => Promise<IAppWebsocket>;
-```
