--- docs/tryorama.iconductor.installagentsapps.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IConductor](./tryorama.iconductor.md) &gt; [installAgentsApps](./tryorama.iconductor.installagentsapps.md)
-
-## IConductor.installAgentsApps property
-
-**Signature:**
-
-```typescript
-installAgentsApps: (options: AgentsAppsOptions) => Promise<AppInfo[]>;
-```
