--- docs/tryorama.iconductor.installapp.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IConductor](./tryorama.iconductor.md) &gt; [installApp](./tryorama.iconductor.installapp.md)
-
-## IConductor.installApp property
-
-**Signature:**
-
-```typescript
-installApp: (appBundleSource: AppBundleSource, options?: AppOptions) => Promise<AppInfo>;
-```
