--- docs/tryorama.iconductor.shutdown.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IConductor](./tryorama.iconductor.md) &gt; [shutDown](./tryorama.iconductor.shutdown.md)
-
-## IConductor.shutDown property
-
-**Signature:**
-
-```typescript
-shutDown: () => Promise<number | null>;
-```
