--- docs/tryorama.iconductor.startup.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IConductor](./tryorama.iconductor.md) &gt; [startUp](./tryorama.iconductor.startup.md)
-
-## IConductor.startUp property
-
-**Signature:**
-
-```typescript
-startUp: () => Promise<void | null>;
-```
