--- docs/tryorama.iconductorcell.cellid.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IConductorCell](./tryorama.iconductorcell.md) &gt; [cellId](./tryorama.iconductorcell.cellid.md)
-
-## IConductorCell.cellId property
-
-**Signature:**
-
-```typescript
-cellId: CellId;
-```
