--- docs/tryorama.iconductorcell.conductor.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IConductorCell](./tryorama.iconductorcell.md) &gt; [conductor](./tryorama.iconductorcell.conductor.md)
-
-## IConductorCell.conductor property
-
-**Signature:**
-
-```typescript
-conductor: IConductor;
-```
