--- docs/tryorama.iplayer.appws.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IPlayer](./tryorama.iplayer.md) &gt; [appWs](./tryorama.iplayer.appws.md)
-
-## IPlayer.appWs property
-
-**Signature:**
-
-```typescript
-appWs: AppWebsocket | IAppWebsocket;
-```
