--- docs/tryorama.iplayer.conductor.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IPlayer](./tryorama.iplayer.md) &gt; [conductor](./tryorama.iplayer.conductor.md)
-
-## IPlayer.conductor property
-
-**Signature:**
-
-```typescript
-conductor: IConductor;
-```
