--- docs/tryorama.iplayer.md
@@ -1,73 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [IPlayer](./tryorama.iplayer.md)
-
-## IPlayer interface
-
-Combines an agent hApp with the conductor they belong to.
-
-**Signature:**
-
-```typescript
-export interface IPlayer extends AgentApp 
-```
-**Extends:** [AgentApp](./tryorama.agentapp.md)
-
-## Properties
-
-<table><thead><tr><th>
-
-Property
-
-
-</th><th>
-
-Modifiers
-
-
-</th><th>
-
-Type
-
-
-</th><th>
-
-Description
-
-
-</th></tr></thead>
-<tbody><tr><td>
-
-[appWs](./tryorama.iplayer.appws.md)
-
-
-</td><td>
-
-
-</td><td>
-
-AppWebsocket \| [IAppWebsocket](./tryorama.iappwebsocket.md)
-
-
-</td><td>
-
-
-</td></tr>
-<tr><td>
-
-[conductor](./tryorama.iplayer.conductor.md)
-
-
-</td><td>
-
-
-</td><td>
-
-[IConductor](./tryorama.iconductor.md)
-
-
-</td><td>
-
-
-</td></tr>
-</tbody></table>
