+++ docs/tryorama.player.appws.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@holochain/tryorama](./tryorama.md) &gt; [Player](./tryorama.player.md) &gt; [appWs](./tryorama.player.appws.md)
+
+## Player.appWs property
+
+**Signature:**
+
+```typescript
+appWs: AppWebsocket;
+```
