--- ts/src/local/index.ts
@@ -1,2 +1,0 @@
-export * from "./conductor.js";
-export * from "./scenario.js";
