+++ ts/test/index.ts
@@ -1,1 +1,2 @@
-import "./local/index.js";
+export * from "./conductor.js";
+export * from "./scenario.js";
